+++ tests/testthat/test-qlq_c30.R
@@ -12,14 +12,16 @@ test_that("qlq_c30: Error if both 'iprefix' and 'items' given", {
 })
 
 
-## These tests are causing problems with CRAN checks.
-## I cannot reproduce, so I am commenting out these tests for now.
+## These tests fail with 2023-10-17 addition to qlq_c30 that produces a
+## message about the re-naming of QLQTOTAL to C30SUMMARY.
+## I am commenting out these tests for now.  I will use again after awhile when I
+## delete the warning msg in qlq_c30.
 
-test_that("qlq_c30: Expect no error (silent) when correct input given", {
-  expect_silent(qlq_c30(df = dat, iprefix = "q"))
-  expect_silent(qlq_c30(df = dat, items = 1:30))
-  expect_silent(qlq_c30(df = dat, items = 1:30, keepNvalid = TRUE))
-})
+# test_that("qlq_c30: Expect no error (silent) when correct input given", {
+#   expect_silent(qlq_c30(df = dat, iprefix = "q"))
+#   expect_silent(qlq_c30(df = dat, items = 1:30))
+#   expect_silent(qlq_c30(df = dat, items = 1:30, keepNvalid = TRUE))
+# })